-Moved translation code to asset manager -Removed animon class, and renamed key_object to animon -Removed event_refs -Removed test_play -Changed playing to feeding and wrote code for feeding play -Fixed some compiler warnings

+++ src/events/event_manager.cpp
@@ -173,7 +173,7 @@ namespace sun_magic {
 	}
 
 	void EventManager::Update() {
-		for (int i = 0; i < events_.size(); ++i) {
+		for (size_t i = 0; i < events_.size(); ++i) {
 			Event event = events_[i];
 
 			EventToFocusToListenerSetMap::iterator event_iter = eventfocus__listener_map_.find(event.type);
